Bugfix/replace admin with something else

+++ src/main/java/cs544_2020_01_light_attendanceproject/controller/UserController.java
@@ -1,48 +1,53 @@
 package cs544_2020_01_light_attendanceproject.controller;
 
 import cs544_2020_01_light_attendanceproject.domain.User;
+import cs544_2020_01_light_attendanceproject.exceptions.AdminsCannotDeleteThemselvesException;
 import cs544_2020_01_light_attendanceproject.exceptions.UserNotFoundException;
 import cs544_2020_01_light_attendanceproject.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 
 @RestController
 @Secured("ROLE_ADMIN")
-@RequestMapping("/admin")
-public class AdminController {
+@RequestMapping("/user")
+public class UserController {
     private UserService userService;
 
     @Autowired
-    public AdminController(UserService userService) {
+    public UserController(UserService userService) {
         this.userService = userService;
     }
 
-    @PostMapping("/user")
+    @PostMapping("/")
     @ResponseStatus(HttpStatus.CREATED)
     public User newUser(@RequestBody @Valid User user) {
         return userService.registerNewUserAccount(user);
     }
 
-    @GetMapping("/user")
+    @GetMapping("/")
     public Iterable<User> all() {
         return userService.listUsers();
     }
 
-    @GetMapping("/user/{username}")
+    @GetMapping("/{username}")
     public User one(@PathVariable String username) {
         return userService.findOneUser(username).orElseThrow(() -> new UserNotFoundException(username));
     }
 
-    @DeleteMapping("/user/{username}")
+    @DeleteMapping("/{username}")
     public void deleteUser(@PathVariable String username) {
+        String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(username.equals(currentUsername))
+            throw new AdminsCannotDeleteThemselvesException(username);
         userService.deleteByUsername(username);
     }
 
-    @PutMapping("/user/{username}")
+    @PutMapping("/{username}")
     public User replaceUser(@RequestBody @Valid User newUser, @PathVariable String username) {
         return userService.replaceUser(newUser, username);
     }
